break some stuff out of lib.rs
